integrated ml-common and fixed import loops
