+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "common"]
+	path = common
+	url = https://github.com/ml4oncology/ml-common
